fix instrumentation test
